requirements.txt
pyserial
viam-sdk==0.49.0
viam-sdk[mlmodel]

typing-extensions
numpy